+++ spec/firmware.rb
@@ -107,6 +107,9 @@ module Firmware
     use_device 'PIC18F25K50'
     use_file Xc8DistDir + 'TestXC8.hex'
     use_symbols RPicSim::Xc8SymFile.new(Xc8DistDir + 'TestXC8.sym')
+
+    def_symbol :postDefinedUnknown, 45
+    def_symbol :postDefinedRam, 46, :ram
   end
 
   class Variables < RPicSim::Sim